dtype

+++ train_video.py
@@ -217,7 +217,7 @@ def main(args):
                 video=batch["video"].permute(0, 2, 1, 3, 4).contiguous().to(device)
                 text=batch["text"]
                 tokenized_text=batch["token"]['input_ids'].to(device)
-                encoder_attention_mask=batch["token"]["attention_mask"].to(device)
+                encoder_attention_mask=batch["token"]["attention_mask"].to(device).float()
                 
                 if e==start_epoch and b==0:
                     accelerator.print("video ",video.size(),video.dtype,video.device)
